financial-coach-app/requirements.txt
# Flask and web framework dependencies
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
python-dotenv==1.0.0

# Machine Learning and AI dependencies
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
joblib==1.3.2

# Database dependencies
SQLAlchemy==2.0.20

# API and validation dependencies
marshmallow==3.20.1
Werkzeug==2.3.7

# Optional: For enhanced ML capabilities
scipy==1.11.2

# Development and testing dependencies (optional)
pytest==7.4.0
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0

# Production server (optional)
gunicorn==21.2.0